default-applications: Use more common icon names for Office documents
"x-office-document" and "x-office-spreadsheet" are more common MIME
icon names than their counterparts without the "x-" prefix.
The latter seems not to even be available in "MATE" icon theme, but
only in "Faenza" icon theme.  The "x-"-prefixed variant on the other
hand is available even in theme providing the unprefixed ones, meaning
it will sill work with these.

On my machine with not many themes, but only standard-ish ones:
```console
$ find /usr/share/icons/ -name office-document.* | wc -l
8
$ find /usr/share/icons/ -name x-office-document.* | wc -l
54
```

Thus, using the prefixed variant seems like a better choice.