windows: Fix UI accessibility (#704)
Use real frames for a semantic UI so GTK and screen readers can set up
and figure out accessibility relationships between the frame label and
the content automatically.

Also properly add relationships between the movement keys descriptive
label and the key radio items so a screen reader can pick it up.  As a
bonus, it also prevents Orca from erroneously announcing the label when
the window comes up because it thinks this long orphaned label is
probably a general description for the window.

Fixes #703.