keybindings: Do not present junk shortcut for headers
For screen reader's sake, make the shortcut empty when hidden,
otherwise a more or less random value is presented.  The actual value
is whatever the render was last used with, remembering cell renderers
are reused for every row.

This could also be fixed in the accessibility implementation side, but
we rely on GTK's, so it's easier to clean it up here than modify GTK or
subclass it further.