font-viewer: Bring the window to the front when triggering the app
If the app is already open, we need to tell the desktop that the window
has been triggered, so it can either bring it to front, or notify the
user the window claims attention.

This fixes the cases where the font viewer is already open and gets
triggered again, e.g. by opening a font or launching it directly.